chore: set gas fee in TxSummaryView
